Rename setup.py in api
