+++ gs/backend/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 
+from gs.backend.api.backend_setup import setup_middlewares, setup_routes
 from gs.backend.api.lifespan import lifespan
-from gs.backend.api.setup import setup_middlewares, setup_routes
 
 app = FastAPI(lifespan=lifespan)
 setup_routes(app)
